# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: TLOZ Data API
  version: 2.0.0-draft
  description: |
    Document-first API for TLOZ Projects, Missions and Inventory.
    The v1 routes remain available temporarily as a compatibility bridge.

    Authentication is via Bearer token (agent API key) or browser session cookie.
    If both are present, the Bearer token takes priority.
  contact:
    name: TLOZ
  license:
    name: Proprietary
    identifier: LicenseRef-Zivelo-Proprietary
servers:
  - url: /api/v1
    description: Legacy v1 compatibility API
security:
  - bearerToken: []
  - sessionCookie: []
...
paths:
...
  /users/{userId}:
    get:
      operationId: getUser
      summary: Get a user by ID
      tags: [Users]
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: User found.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: authjs.session-token
      description: Existing NextAuth browser session. Production may use the secure cookie prefix.
    bearerToken:
      type: http
      scheme: bearer
      bearerFormat: tloz_
      description: |
        New agent API keys use `tloz_` followed by 64 hex characters.
        Legacy `zaf_` keys remain valid during the v1 grace release.
        To authenticate, include the header:
        ```
        Authorization: Bearer tloz_<64-hex-chars>
        ```
        If both Bearer token and session cookie are present, the Bearer token takes priority.
  parameters:
...
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
...
  responses:
...
    Unauthorized:
      description: An authenticated session or valid API key is required.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Forbidden:
      description: The authenticated actor is not permitted to perform this operation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    NotFound:
      description: The requested resource does not exist.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    InternalError:
      description: Unexpected server error. Details are recorded server-side.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
  schemas:
...
    ApiError:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, requestId]
          properties:
            code:
              type: string
              enum: [INVALID_REQUEST, UNAUTHORIZED, FORBIDDEN, NOT_FOUND, INTERNAL_ERROR, DOCUMENT_INVALID, DOCUMENT_NOT_FOUND, DOCUMENT_REVISION_CONFLICT, PRECONDITION_REQUIRED, INVALID_ATTACHMENT_MANIFEST, ATTACHMENT_STORAGE_NOT_CONFIGURED, ATTACHMENT_STORAGE_ERROR, attachment_object_missing, attachment_object_invalid, batch_superseded]
            message:
              type: string
            fields:
              type: object
              description: Field-specific guidance for INVALID_REQUEST responses.
              additionalProperties:
                type: string
            requestId:
              type: string

    # ── Users ────────────────────────────────────────────────────────
    User:
      type: object
      additionalProperties: false
      description: Operational profile. Email is omitted for agent:reader; passwords, sessions, API keys, and other authentication secrets are never exposed.
      required: [id, name, username, role, type, avatarUrl]
      properties:
        id:
          type: string
        name:
          type: string
        username:
          type: string
        email:
          type: string
          format: email
          description: Omitted from responses to agent:reader.
        role:
          $ref: "#/components/schemas/UserRole"
        type:
          type: string
          enum: [human, agent]
        avatarUrl:
          type: string
          format: uri
        theme:
          type: string
          enum: [system, light, dark]
...
    UserRole:
      type: string
      enum: [Platform Owner, Full Stack Developer, agent:operative, agent:reader]

    # ── Projects ─────────────────────────────────────────────────────
...